admin/requirements.txt
django==4.2.4
django-environ==0.10.0
django-cors-headers==4.2.0
django-extensions==3.2.3
psycopg2-binary==2.9.7
django-postgres-extra==2.0.8
django-split-settings==1.2.0
django-admin-rangefilter==0.10.0
django-admin-autocomplete-all==0.6.6

uWSGI==2.0.22

requests==2.31.0
PyJWT==2.8.0

sentry-sdk==1.42.0
